TensorIteratorReduce: Avoid tensor operations in parallel_for
Ref gh-56794

The two pass reduction calls `copy_` and `select` inside a parallel region. The
`copy_` can just be moved outside of the parallel region, but avoiding the
`select` call is more complicated because it's needed to construct the
`TensorIterator`. Instead, I factor out a `serial_for_each` free-function that
just takes pointers and strides. Then manually advance the pointer to the
thread-specific slice of data.

ghstack-source-id: 754370e06105b3637733cb99a954400baa18228a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58655